DRAFT: init RemoteDatabase w/ fork_block_number
Draft status: this is okay for satisfying tests that query data from the
fork block number passed in via the fork node config, but for tests that
then query blocks PRIOR to that fork block number, this obviously
doesn't help at all. but the Database API doesn't offer a way to pass in
a particular block number as a parameter to eg basic(), so what will we
do? perhaps simply have a "setBlockNumber()" method inherent to the
RemoteDatabase struct, which will allow the client to set the block
number just prior to calling eg basic()? seems pretty hacky, but seems
like it'll work.